perf(functions): Improve weld performance in extreme cases
- Adds compactPrimitive and compactAttribute functions
- Ensure that welding copies no more of a primitive's vertices than necessary
